:recycle: issue #2 DDDアーキテクチャに則り、`CustomProgressBar`を各レイヤーに責務分離
